feat: add tracked models and operations with locking mechanisms for invoice and payment processing
