+++ src/v4vapp_backend_v2/actions/tracked_all.py
@@ -1,0 +1,33 @@
+from typing import Any, Union
+
+from v4vapp_backend_v2.hive_models.op_all import OpAny, op_any
+from v4vapp_backend_v2.models.invoice_models import Invoice
+from v4vapp_backend_v2.models.payment_models import Payment
+
+TrackedAny = Union[OpAny, Invoice, Payment]
+
+
+def tracked_any(tracked: dict[str, Any]) -> TrackedAny:
+    """
+    Check if the tracked object is of type OpAny, Invoice, or Payment.
+
+    :param tracked: The object to check.
+    :return: True if the object is of type OpAny, Invoice, or Payment, False otherwise.
+    """
+    if isinstance(tracked, OpAny):
+        return tracked
+    elif isinstance(tracked, Invoice) or isinstance(tracked, Payment):
+        return tracked
+
+    if tracked.get("type", None):
+        try:
+            return op_any(tracked)
+        except Exception as e:
+            raise ValueError(f"Invalid tracked object: {e}")
+
+    if tracked.get("r_hash", None):
+        return Invoice.model_validate(tracked)
+    if tracked.get("payment_hash", None):
+        return Payment.model_validate(tracked)
+
+    raise ValueError("Invalid tracked object")
