+++ src/v4vapp_backend_v2/actions/tracked_models.py
@@ -1,0 +1,94 @@
+from typing import ClassVar
+
+from pydantic import BaseModel, Field
+from pymongo.results import UpdateResult
+
+from v4vapp_backend_v2.database.db import MongoDBClient
+
+
+class TrackedBaseModel(BaseModel):
+    locked: bool = Field(
+        default=False,
+        description="Flag to indicate if the operation is locked or being processed",
+        exclude=False,
+    )
+
+    db_client: ClassVar[MongoDBClient | None] = None
+
+    def __init__(self, **data):
+        """
+        Initialize the TrackedBaseModel with the provided data.
+
+        :param data: The data to initialize the model with.
+        """
+        super().__init__(**data)
+        self.locked = data.get("locked", False)
+
+    @property
+    def collection(self) -> str:
+        """
+        Returns the name of the collection associated with this model.
+
+        This method should be overridden in subclasses to provide the
+        specific collection name.
+
+        Returns:
+            str: The name of the collection.
+        """
+        raise NotImplementedError("Subclasses must implement this method.")
+
+    @property
+    def group_id_query(self) -> dict:
+        """
+        Returns the query used to identify the group ID in the database.
+
+        This method should be overridden in subclasses to provide the
+        specific query for the group ID.
+
+        Returns:
+            dict: The query used to identify the group ID.
+        """
+        raise NotImplementedError("Subclasses must implement this method.")
+
+    async def lock_op(self) -> UpdateResult | None:
+        """
+        Locks the operation to prevent concurrent processing.
+
+        This method sets the `locked` attribute to True, indicating that
+        the operation is currently being processed and should not be
+        modified or accessed by other threads or processes.
+
+        Returns:
+            None
+        """
+        self.locked = True
+        if self.db_client:
+            ans = await self.db_client.update_one(
+                collection_name=self.collection,
+                query=self.group_id_query,
+                update={"locked": self.locked},
+            )
+            return ans
+        return None
+
+    async def unlock_op(self) -> UpdateResult | None:
+        """
+        Unlocks the operation to allow concurrent processing.
+
+        This method sets the `locked` attribute to False, indicating that
+        the operation is no longer being processed and can be modified
+        or accessed by other threads or processes.
+
+        Returns:
+            None
+        """
+        self.locked = False
+        if self.db_client:
+            # Remember my update_one already has the $set
+            ans = await self.db_client.update_one(
+                collection_name=self.collection,
+                query=self.group_id_query,
+                update={"$unset": {"locked": ""}},
+            )
+            return ans
+        return None
