+++ src/v4vapp_backend_v2/hive_models/op_base.py
@@ -1,168 +1,28 @@
 import re
 from asyncio import get_event_loop
 from datetime import datetime, timezone
-from enum import StrEnum, auto
 from typing import Any, ClassVar, Dict, List
 
 from nectar import Hive
-from pydantic import BaseModel, Field, computed_field
+from pydantic import Field, computed_field
 
-from v4vapp_backend_v2.accounting.account_type import AccountAny
+from v4vapp_backend_v2.actions.tracked_models import TrackedBaseModel
 from v4vapp_backend_v2.config.setup import logger
 from v4vapp_backend_v2.database.db import MongoDBClient
 from v4vapp_backend_v2.helpers.crypto_conversion import CryptoConversion
 from v4vapp_backend_v2.helpers.crypto_prices import AllQuotes, QuoteResponse
 from v4vapp_backend_v2.helpers.general_purpose_funcs import format_time_delta, snake_case
 from v4vapp_backend_v2.hive_models.custom_json_data import all_custom_json_ids, custom_json_test_id
-from v4vapp_backend_v2.hive_models.real_virtual_ops import HIVE_REAL_OPS, HIVE_VIRTUAL_OPS
-
-# This list needs to be synced with op_all.py
-OP_TRACKED = [
-    "custom_json",
-    "transfer",
-    "account_witness_vote",
-    "producer_reward",
-    "fill_order",
-    "limit_order_create",
-    "update_proposal_votes",
-    "account_update2",
-    "fill_recurrent_transfer",
-    "recurrent_transfer",
-]
-
-# I don't think I need the recurrent_transfer_cancel op because it doesn't affect me.
+from v4vapp_backend_v2.hive_models.op_base_extras import (
+    OP_TRACKED,
+    HiveExp,
+    OpLogData,
+    OpRealm,
+    op_realm,
+)
 
 
-class OpRealm(StrEnum):
-    REAL = auto()
-    VIRTUAL = auto()
-    MARKER = auto()
-
-
-class HiveExp(StrEnum):
-    HiveHub = "https://hivehub.dev/{prefix_path}"
-    HiveScanInfo = "https://hivescan.info/{prefix_path}"
-    # HiveBlockExplorer = "https://hiveblockexplorer.com/{prefix_path}"
-    # HiveExplorer = "https://hivexplorer.com/{prefix_path}"
-
-
-def get_hive_block_explorer_link(
-    trx_id: str,
-    block_explorer: HiveExp = HiveExp.HiveHub,
-    markdown: bool = False,
-    block_num: int = 0,
-    op_in_trx: int = 0,
-    realm: OpRealm = OpRealm.REAL,
-    # any_op: OpBase | None = None,
-) -> str:
-    """
-    Generate a Hive blockchain explorer URL for a given transaction ID.
-
-    Args:
-        trx_id (str): The transaction ID to include in the URL
-        block_explorer (HiveExp): The blockchain explorer to use (defaults to HiveHub)
-
-    Returns:
-        str: The complete URL with the transaction ID inserted
-    """
-
-    if trx_id and not (block_num and op_in_trx):
-        path = f"{trx_id}"
-        prefix = "tx/"
-    elif trx_id == "0000000000000000000000000000000000000000" and block_num:
-        op_in_trx = op_in_trx if op_in_trx else 1
-        prefix = f"{block_num}/"
-        path = f"{trx_id}/{op_in_trx}"
-    elif trx_id and block_num and op_in_trx and realm == OpRealm.VIRTUAL:
-        path = f"{block_num}/{trx_id}/{op_in_trx}"
-        prefix = "tx/"
-    elif trx_id and op_in_trx and realm == OpRealm.REAL:
-        if op_in_trx > 1:
-            path = f"{trx_id}/{op_in_trx}"
-        else:
-            path = f"{trx_id}"
-        prefix = "tx/"
-    elif not trx_id and block_num:
-        path = f"{block_num}"
-        prefix = "b/"
-
-    if block_explorer == HiveExp.HiveScanInfo:
-        if prefix == "tx/":
-            prefix = "transaction/"
-        elif prefix == "b/":
-            prefix = "block/"
-
-    prefix_path = f"{prefix}{path}"
-
-    link_html = block_explorer.value.format(prefix_path=prefix_path)
-    if not markdown:
-        return link_html
-    markdown_link = f"[{block_explorer.name}]({link_html})"
-    return markdown_link
-
-
-def op_realm(op_type: str):
-    """
-    Determines the operational realm based on the provided operation type.
-
-    Args:
-        op_type (str): The type of operation to evaluate. Can be one of the
-                       predefined operation types or None.
-
-    Returns:
-        OpRealm: The corresponding operational realm, which can be one of the
-                 following:
-                 - OpRealm.VIRTUAL: If the operation type is in HIVE_VIRTUAL_OPS.
-                 - OpRealm.REAL: If the operation type is in HIVE_REAL_OPS.
-                 - OpRealm.MARKER: If the operation type is "block_marker".
-                 - None: If the operation type is None or does not match any
-                         predefined types.
-    """
-    if op_type is not None:
-        if op_type in HIVE_VIRTUAL_OPS:
-            return OpRealm.VIRTUAL
-        elif op_type in HIVE_REAL_OPS:
-            return OpRealm.REAL
-        elif op_type == "block_marker":
-            return OpRealm.MARKER
-    return None
-
-
-class OpLogData(BaseModel):
-    """
-    OpLogData is a Pydantic model that represents the structure of log data.
-
-    Attributes:
-        log (str): The main log message.
-        notification (str): A notification message associated with the log.
-        log_extra (Dict[str, Any]): Additional data or metadata related to the log.
-    """
-
-    log: str
-    notification: str
-    log_extra: Dict[str, Any]
-
-
-class WatchPair(BaseModel):
-    """
-    WatchPair is a Pydantic model that represents a pair of accounts to watch.
-
-    Attributes:
-        from_account (str): The account from which the operation originates.
-        to_account (str): The account to which the operation is directed.
-    """
-
-    from_account: str | None = None
-    to_account: str | None = None
-    ledger_debit: AccountAny | None = None
-    ledger_credit: AccountAny | None = None
-    ledger_fee: AccountAny | None = None
-
-    def __init__(self, **data):
-        super().__init__(**data)
-
-
-class OpBase(BaseModel):
+class OpBase(TrackedBaseModel):
     """
     OpBase is a base model representing a Hive blockchain operation. It provides attributes
     and methods to handle both real and virtual operations, along with logging and notification
@@ -217,6 +77,7 @@ class OpBase(BaseModel):
             an HTML link or a markdown link based on the `markdown` parameter.
     """
 
+    # Hive operation attributes
     realm: OpRealm = Field(
         default=OpRealm.REAL,
         description=(
@@ -233,6 +94,7 @@ class OpBase(BaseModel):
         default=[], description="List of extensions associated with the operation"
     )
 
+    # Raw operations as delivered by Nectar
     raw_op: dict[str, Any] = Field(
         default={}, description="Raw operation data from the blockchain", exclude=True
     )
@@ -300,6 +162,19 @@ class OpBase(BaseModel):
         }
         # special case for OpRealm.MARKER (Overrides this default)
         return ans
+
+    @property
+    def collection(self) -> str:
+        """
+        Returns the name of the collection associated with this model.
+
+        This method is used to determine where the operation data will be stored
+        in the database.
+
+        Returns:
+            str: The name of the collection.
+        """
+        return "hive_ops"
 
     @property
     def group_id(self) -> str:
@@ -477,6 +352,44 @@ class OpBase(BaseModel):
             await all_quotes.get_all_quotes()
             cls.last_quote = all_quotes.quote
 
+    async def lock_op(self) -> None:
+        """
+        Locks the operation to prevent concurrent processing.
+
+        This method sets the `_locked` attribute to True, indicating that
+        the operation is currently being processed and should not be
+        modified or accessed by other threads or processes.
+
+        Returns:
+            None
+        """
+        self._locked = True
+        if self.db_client:
+            await self.db_client.update_one(
+                collection="hive_ops",
+                query=self.group_id_query,
+                update={"$set": {"_locked": self._locked}},
+            )
+
+    async def unlock_op(self) -> None:
+        """
+        Unlocks the operation to allow concurrent processing.
+
+        This method sets the `_locked` attribute to False, indicating that
+        the operation is no longer being processed and can be modified
+        or accessed by other threads or processes.
+
+        Returns:
+            None
+        """
+        self._locked = False
+        if self.db_client:
+            await self.db_client.update_one(
+                collection="hive_ops",
+                query=self.group_id_query,
+                update={"$set": {"_locked": self._locked}},
+            )
+
     async def update_quote_conv(self, quote: QuoteResponse | None = None) -> None:
         """
         Asynchronously updates the last quote for the class.
@@ -500,7 +413,7 @@ class OpBase(BaseModel):
         """
         Updates the conversion for the transaction.
 
-        If the subclass has a `conv` object, update it with the lastest quote.
+        If the subclass has a `conv` object, update it with the latest quote.
         If a quote is provided, it sets the conversion to the provided quote.
         If no quote is provided, it uses the last quote to set the conversion.
 
@@ -574,7 +487,7 @@ class OpBase(BaseModel):
     @property
     def lightning_memo(self) -> str:
         """
-        Removes and shortens a lightning invoice from a memo for outpu.
+        Removes and shortens a lightning invoice from a memo for output.
 
         Returns:
             str: The shortened memo string.
